feat: wasm32 bridge + fix size_t overflow in capi (#4)
* fix: correct size_t overflow in gvRenderData callers (capi)

The C wrapper declared `unsigned int length` but passed `&length` to
gvRenderData which expects `size_t *` — an 8-byte write into a 4-byte
stack slot on 64-bit targets, corrupting adjacent locals. The resulting
heap pointer drift caused segfault/SIGABRT inside gvFreeRenderData().

Fix both call sites to use `size_t` locals.


* feat: add wasm-bindgen bridge so graphviz-anywhere compiles for wasm32

When the target is `wasm32-unknown-unknown` we skip all native linking
and route `GraphvizContext::render` through a JS extern:

    globalThis.__graphviz_anywhere_render(dot, engine, format) -> string

Throwing a JS Error surfaces as GraphvizError::RenderFailed. The native
path is unchanged.

- Gate ffi module and all CStr/ptr-using code behind cfg(not(wasm32)).
- Add packages/rust/src/wasm.rs with the JS extern + render shim.
- build.rs: early-return on CARGO_CFG_TARGET_ARCH == "wasm32"; no prebuilt
  or env-var lookup required.
- Document the bridge contract + example wire-up via
  @kookyleo/graphviz-anywhere-web in packages/rust/README.md.

This lets downstream crates like plantuml-little-web compile once for
wasm and reuse the same API shape they use natively.


---------

+++ capi/graphviz_api.c
@@ -87,7 +87,7 @@ GV_API gv_error_t gv_render(gv_context_t *ctx,
 
     /* Render to memory buffer */
     char *result = NULL;
-    unsigned int length = 0;
+    size_t length = 0;
     rc = gvRenderData(ctx->gvc, g, format, &result, &length);
     if (rc != 0) {
         gvFreeLayout(ctx->gvc, g);
@@ -96,7 +96,7 @@ GV_API gv_error_t gv_render(gv_context_t *ctx,
     }
 
     *out_data = result;
-    *out_length = (size_t)length;
+    *out_length = length;
 
     /* Cleanup graph and layout, but not the render data */
     gvFreeLayout(ctx->gvc, g);
@@ -257,7 +257,7 @@ GV_API gv_error_t gv_render_formats(gv_context_t *ctx,
 
                         /* Render to this format */
                         char *result = NULL;
-                        unsigned int result_len = 0;
+                        size_t result_len = 0;
                         int render_rc = gvRenderData(ctx->gvc, g, fmt_list[fmt_count],
                                                      &result, &result_len);
                         result_list[fmt_count] = (render_rc == 0) ? result : NULL;
